requirements.txt
# Core ML/AI
torch>=2.0.0
sentence-transformers>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0

# Vector Stores
faiss-cpu>=1.7.4
# pinecone-client>=2.2.0  # Uncomment if using Pinecone

# Document Loading
pypdf>=4.0.0
pyyaml>=6.0

# API (optional, for production)
# fastapi>=0.100.0
# uvicorn[standard]>=0.23.0
# python-multipart>=0.0.6
# pydantic>=2.0.0
# pydantic-settings>=2.0.0

# LLM Integration (optional)
# openai>=1.0.0
# anthropic>=0.3.0

# Monitoring (optional)
# prometheus-client>=0.17.0
# slowapi>=0.1.9  # Rate limiting

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Development
ipython>=8.14.0
jupyter>=1.0.0
